testing servicios

+++ tp5/src/main/java/com/example/tp5/repositories/EmpleadoRepository.java
@@ -16,7 +16,7 @@ import java.util.Optional;
 public interface EmpleadoRepository extends JpaRepository<Empleado, Long> {
     Optional<Empleado> findByEmail(String email);
     List<Empleado> findByDepartamento(Departamento departamento);
-    List<Empleado> findBySalarioBetween(BigDecimal salarioMin, BigDecimal salarioMax);
+    List<Empleado> findBySalarioBetween(Double salarioMin, Double salarioMax);
     List<Empleado> findByFechaContratacionAfter(LocalDate fecha);
 
     @Query("SELECT e FROM Empleado e WHERE e.departamento.nombre = :nombreDepartamento")
